Move document-capture to new packages directory
