+++ app/javascript/packages/document-capture/components/acuant-capture-canvas.jsx
