+++ app/javascript/packages/document-capture/components/acuant-capture.jsx
