+++ app/javascript/packages/document-capture/components/button.jsx
