+++ app/javascript/packages/document-capture/components/document-capture.jsx
