+++ app/javascript/packages/document-capture/components/documents-step.jsx
