+++ app/javascript/packages/document-capture/components/file-input.jsx
