+++ app/javascript/packages/document-capture/components/full-screen.jsx
