+++ app/javascript/packages/document-capture/components/image.jsx
