+++ app/javascript/packages/document-capture/components/mobile-intro-step.jsx
