+++ app/javascript/packages/document-capture/components/selfie-step.jsx
