+++ app/javascript/packages/document-capture/components/submission-complete.jsx
