+++ app/javascript/packages/document-capture/components/submission.jsx
