+++ app/javascript/packages/document-capture/context/asset.js
