+++ app/javascript/packages/document-capture/context/device.js
