+++ app/javascript/packages/document-capture/context/i18n.js
