+++ app/javascript/packages/document-capture/context/upload.js
