+++ app/javascript/packages/document-capture/hooks/use-async.js
