+++ app/javascript/packages/document-capture/hooks/use-i18n.js
