+++ app/javascript/packages/document-capture/hooks/use-if-still-mounted.js
