+++ app/javascript/packages/document-capture/hooks/use-instance-id.js
