+++ app/javascript/packages/document-capture/services/upload.js
