+++ app/javascript/packs/document-capture.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { render } from 'react-dom';
-import DocumentCapture from '../app/document-capture/components/document-capture';
-import AssetContext from '../app/document-capture/context/asset';
-import I18nContext from '../app/document-capture/context/i18n';
-import DeviceContext from '../app/document-capture/context/device';
-import { Provider as AcuantProvider } from '../app/document-capture/context/acuant';
+import DocumentCapture from '../packages/document-capture/components/document-capture';
+import AssetContext from '../packages/document-capture/context/asset';
+import I18nContext from '../packages/document-capture/context/i18n';
+import DeviceContext from '../packages/document-capture/context/device';
+import { Provider as AcuantProvider } from '../packages/document-capture/context/acuant';
 
 const { I18n: i18n, assets } = window.LoginGov;
 
@@ -12,7 +12,7 @@ function getMetaContent(name) {
   return document.querySelector(`meta[name="${name}"]`)?.content ?? null;
 }
 
-/** @type {import('../app/document-capture/context/device').DeviceContext} */
+/** @type {import('../packages/document-capture/context/device').DeviceContext} */
 const device = {
   isMobile:
     'mediaDevices' in window.navigator &&
