+++ spec/javascripts/app/document-capture/components/button-spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import sinon from 'sinon';
 import userEvent from '@testing-library/user-event';
 import render from '../../../support/render';
-import Button from '../../../../../app/javascript/app/document-capture/components/button';
+import Button from '../../../../../app/javascript/packages/document-capture/components/button';
 
 describe('document-capture/components/button', () => {
   it('renders with default props', () => {
