+++ spec/javascripts/app/document-capture/context/acuant-spec.jsx
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 import render from '../../../support/render';
 import AcuantContext, {
   Provider as AcuantContextProvider,
-} from '../../../../../app/javascript/app/document-capture/context/acuant';
+} from '../../../../../app/javascript/packages/document-capture/context/acuant';
 
 describe('document-capture/context/acuant', () => {
   afterEach(() => {
