+++ spec/javascripts/app/document-capture/context/asset-spec.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import render from '../../../support/render';
-import AssetContext from '../../../../../app/javascript/app/document-capture/context/asset';
+import AssetContext from '../../../../../app/javascript/packages/document-capture/context/asset';
 
 describe('document-capture/context/asset', () => {
   const ContextValue = () => JSON.stringify(useContext(AssetContext));
