+++ spec/javascripts/app/document-capture/context/i18n-spec.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import render from '../../../support/render';
-import I18nContext from '../../../../../app/javascript/app/document-capture/context/i18n';
+import I18nContext from '../../../../../app/javascript/packages/document-capture/context/i18n';
 
 describe('document-capture/context/i18n', () => {
   const ContextValue = () => JSON.stringify(useContext(I18nContext));
