+++ spec/javascripts/app/document-capture/hooks/use-async-spec.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import sinon from 'sinon';
 import render from '../../../support/render';
-import useAsync from '../../../../../app/javascript/app/document-capture/hooks/use-async';
-import SuspenseErrorBoundary from '../../../../../app/javascript/app/document-capture/components/suspense-error-boundary';
+import useAsync from '../../../../../app/javascript/packages/document-capture/hooks/use-async';
+import SuspenseErrorBoundary from '../../../../../app/javascript/packages/document-capture/components/suspense-error-boundary';
 
 describe('document-capture/hooks/use-async', () => {
   function Child({ resource }) {
