+++ spec/javascripts/app/document-capture/hooks/use-i18n-spec.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import render from '../../../support/render';
-import I18nContext from '../../../../../app/javascript/app/document-capture/context/i18n';
+import I18nContext from '../../../../../app/javascript/packages/document-capture/context/i18n';
 import useI18n, {
   formatHTML,
-} from '../../../../../app/javascript/app/document-capture/hooks/use-i18n';
+} from '../../../../../app/javascript/packages/document-capture/hooks/use-i18n';
 
 describe('document-capture/hooks/use-i18n', () => {
   describe('formatHTML', () => {
